:hammer: REFACTOR (project): extract ProjectModel declaration
