+++ src/models/Project/ProjectModel.ts
@@ -1,0 +1,10 @@
+import AssetModel from "models/Asset";
+
+interface ProjectModel {
+  title: string;
+  tags: string[];
+  thumbnail: AssetModel;
+  rows: string[];
+}
+
+export default ProjectModel;
