+++ src/models/Project/index.ts
@@ -1,13 +1,5 @@
-import AssetModel from "models/Asset";
-
+import ProjectModel from "./ProjectModel";
 import ProjectQueryModel from "./ProjectQueryModel";
-
-interface ProjectModel {
-  title: string;
-  tags: string[];
-  thumbnail: AssetModel;
-  rows: string[];
-}
 
 export default ProjectModel;
 export { ProjectQueryModel };
